refactor(*): split the ui-helper module into separate modules with clear purposes
