+++ src/app/child-dev-project/attendance/attendance-details/attendance-details.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject, Input, OnInit, ViewChild } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { ConfirmationDialogService } from '../../../core/ui-helper/confirmation-dialog/confirmation-dialog.service';
+import { ConfirmationDialogService } from '../../../core/confirmation-dialog/confirmation-dialog.service';
 import { EntityMapperService } from '../../../core/entity/entity-mapper.service';
 import { AttendanceMonth } from '../model/attendance-month';
