+++ src/app/child-dev-project/children/child-details/child-details.component.spec.ts
@@ -14,7 +14,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { FormBuilder, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { EntityMapperService } from '../../../core/entity/entity-mapper.service';
 import { AlertService } from '../../../core/alerts/alert.service';
-import { ConfirmationDialogService } from '../../../core/ui-helper/confirmation-dialog/confirmation-dialog.service';
+import { ConfirmationDialogService } from '../../../core/confirmation-dialog/confirmation-dialog.service';
 import { DatePipe, Location, PercentPipe } from '@angular/common';
 import { Observable } from 'rxjs';
 import { ChildDetailsComponent } from './child-details.component';
@@ -23,8 +23,8 @@ import { AserComponent } from '../../aser/aser-component/aser.component';
 import { ChildAttendanceComponent } from '../../attendance/child-attendance/child-attendance.component';
 import { NotesComponent } from '../../notes/notes-component/notes.component';
 import { EducationalMaterialComponent } from '../../educational-material/educational-material-component/educational-material.component';
-import { KeysPipe } from '../../../core/ui-helper/keys-pipe/keys.pipe';
-import { EntitySubrecordComponent } from '../../../core/ui-helper/entity-subrecord/entity-subrecord.component';
+import { KeysPipe } from '../../../core/entity-subrecord/keys-pipe/keys.pipe';
+import { EntitySubrecordComponent } from '../../../core/entity-subrecord/entity-subrecord/entity-subrecord.component';
 import { AttendanceDaysComponent } from '../../attendance/attendance-days/attendance-days.component';
 import { AttendanceDayBlockComponent } from '../../attendance/attendance-days/attendance-day-block.component';
 import { ChildrenService } from '../children.service';
