+++ src/app/child-dev-project/notes/notes-manager/notes-manager.component.ts
@@ -7,7 +7,7 @@ import { Note } from '../model/note';
 import { NoteDetailsComponent } from '../note-details/note-details.component';
 import { Subscription } from 'rxjs';
 import { SessionService } from '../../../core/session/session.service';
-import { FilterSelection } from '../../../core/ui-helper/filter-selection/filter-selection';
+import { FilterSelection } from '../../../core/filter/filter-selection/filter-selection';
 import { WarningLevel } from '../../warning-level';
 import { MediaChange, MediaObserver } from '@angular/flex-layout';
