+++ src/app/child-dev-project/schools/school-detail/school-detail.component.spec.ts
@@ -14,7 +14,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
 import { EntityMapperService } from '../../../core/entity/entity-mapper.service';
 import { AlertService } from '../../../core/alerts/alert.service';
-import { ConfirmationDialogService } from '../../../core/ui-helper/confirmation-dialog/confirmation-dialog.service';
+import { ConfirmationDialogService } from '../../../core/confirmation-dialog/confirmation-dialog.service';
 import { Database } from '../../../core/database/database';
 import { MockDatabase } from '../../../core/database/mock-database';
 import { Location } from '@angular/common';
