+++ src/app/core/confirmation-dialog/confirmation-dialog/confirmation-dialog.component.scss
