+++ src/app/core/entity-subrecord/entity-subrecord.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ConfirmationDialogComponent } from './confirmation-dialog/confirmation-dialog.component';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule } from '@angular/material/dialog';
@@ -10,10 +9,10 @@ import { MatSelectModule } from '@angular/material/select';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatSortModule } from '@angular/material/sort';
 import { MatTableModule } from '@angular/material/table';
-import { ConfirmationDialogService } from './confirmation-dialog/confirmation-dialog.service';
 import { EntitySubrecordComponent } from './entity-subrecord/entity-subrecord.component';
 import { KeysPipe } from './keys-pipe/keys.pipe';
 import { MatDatepickerModule } from '@angular/material/datepicker';
+import { ConfirmationDialogModule } from '../confirmation-dialog/confirmation-dialog.module';
 
 @NgModule({
   imports: [
@@ -28,10 +27,9 @@ import { MatDatepickerModule } from '@angular/material/datepicker';
     MatSortModule,
     MatAutocompleteModule,
     MatDatepickerModule,
+    ConfirmationDialogModule,
   ],
-  declarations: [ConfirmationDialogComponent, EntitySubrecordComponent, KeysPipe],
+  declarations: [EntitySubrecordComponent, KeysPipe],
   exports: [EntitySubrecordComponent, KeysPipe],
-  providers: [ConfirmationDialogService],
-  entryComponents: [ConfirmationDialogComponent],
 })
-export class UiHelperModule { }
+export class EntitySubrecordModule { }
