+++ src/app/core/entity-subrecord/entity-subrecord/column-description.ts
