+++ src/app/core/entity-subrecord/entity-subrecord/entity-subrecord.component.scss
