+++ src/app/core/entity-subrecord/entity-subrecord/entity-subrecord.component.ts
@@ -3,7 +3,6 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
-import { ConfirmationDialogService } from '../confirmation-dialog/confirmation-dialog.service';
 import { Entity } from '../../entity/entity';
 import { EntityMapperService } from '../../entity/entity-mapper.service';
 import { ColumnDescription, ColumnDescriptionInputType } from './column-description';
@@ -11,6 +10,7 @@ import { MediaChange, MediaObserver } from '@angular/flex-layout';
 import { Subscription } from 'rxjs';
 import { AlertService } from 'app/core/alerts/alert.service';
 import { FormValidationResult } from './form-validation-result';
+import { ConfirmationDialogService } from '../../confirmation-dialog/confirmation-dialog.service';
 
 
 @Component({
