--- src/app/core/ui-helper/ui-helper.module.spec.ts
@@ -1,13 +1,0 @@
-import { UiHelperModule } from './ui-helper.module';
-
-describe('UiHelperModule', () => {
-  let uiComponentsModule: UiHelperModule;
-
-  beforeEach(() => {
-    uiComponentsModule = new UiHelperModule();
-  });
-
-  it('should create an instance', () => {
-    expect(uiComponentsModule).toBeTruthy();
-  });
-});
